+++ src/1-one/tokenize/compute/normal/02-acronyms.js
@@ -1,3 +1,4 @@
+// do acronyms need to be ASCII?  ... kind of?
 const periodAcronym = /([A-Z]\.)+[A-Z]?,?$/
 const oneLetterAcronym = /^[A-Z]\.,?$/
 const noPeriodAcronym = /[A-Z]{2,}('s|,)?$/