+++ src/1-one/tokenize/methods/01-sentences/02-isSentence.js
@@ -1,6 +1,6 @@
 const isAcronym = /[ .][A-Z]\.? *$/i
 const hasEllipse = /(?:\u2026|\.{2,}) *$/
-const hasLetter = /[a-z0-9\u00C0-\u00FF\u00a9\u00ae\u2000-\u3300\ud000-\udfff]/i
+const hasLetter = /\p{L}/u
 
 /** does this look like a sentence? */
 const isSentence = function (str, abbrevs) {