+++ src/1-one/tokenize/methods/02-terms/02-slashes.js
@@ -1,4 +1,5 @@
-const isSlash = /[a-z] ?\/ ?[a-z]+$/
+const isSlash = /\p{L} ?\/ ?\p{L}+$/u
+
 // 'he / she' should be one word
 const combineSlashes = function (arr) {
   for (let i = 1; i < arr.length - 1; i++) {