+++ changelog.md
@@ -12,6 +12,7 @@ While all _Major_ releases should be reviewed, our only two _large_ releases are
 - **[fix]** - bug with fast-or possessive matches
 - **[new]** - add novel tag detection on world.one.tagSet 
 - **[new]** - proof-of-concept nlp.makeNet() method
+- **[fix]** - better regex unicode support
 -->
 
 #### 14.1.2 [April 2021]